// src/utils/types.ts
export type CGPAscaleType = 4 | 5 | 7;

export type GradeType = "A" | "B" | "C" | "D" | "E" | "F";

export type GPAClassType =
  | ""
  | "First Class"
  | "Second Class (Upper)"
  | "Second Class (Lower)"
  | "Third Class"
  | "Pass"
  | "Probation";

export type RemarkType =
  | ""
  | "Excellent"
  | "Very Good"
  | "Good"
  | "Fair"
  | "Poor";